requirements.txt
# Core dependencies for Voxel ambient art generator
sounddevice>=0.4.6
vosk>=0.3.45
openai>=1.0.0
pygame>=2.5.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# Web interface dependencies
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# Google Cloud dependencies (optional - for Google Cloud image generation)
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0